requirements.txt
-e src/packages/azure_utils
-e src/packages/tfod_utils

cryptography>=2.8
flake8==3.8.4
pytest==6.2.2
azure-storage-blob==12.7.1
azure-storage-file==2.1.0
azureml-sdk==1.35.0
numpy==1.19.3
pandas==1.1.5
Pillow==8.4.0
jellyfish==0.8.2
tensorflow==1.15
matplotlib==3.3.4
opencv-python-headless==4.5.3.56
streamlit